Adjusted changes

+++ backend/src/main/java/com/tixly/ticket/models/request/BusRequest.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class BusRequest {
     private String plateNo;
     private int seatNo;
-    private int companyId;
+    private Long companyId;
     private String busType;
 }
